hotfix/#263: List.of 파라미터로 null이 전해짐에 따른 NullPointerException 수정

+++ src/main/java/com/hanpyeon/academyapi/online/service/lesson/update/LessonImageUpdateHandler.java
@@ -23,9 +23,11 @@ class LessonImageUpdateHandler implements LessonUpdateHandler{
         if (Objects.isNull(updateOnlineLessonInfoCommand.imageSrc())) {
             return;
         }
-        imageService.removeImage(List.of(onlineCourse.getImage()));
-        final Image preImage = loadImage(updateOnlineLessonInfoCommand.imageSrc());
-        onlineCourse.setImage(preImage);
+        if (onlineCourse.getImage() != null) {
+            imageService.removeImage(List.of(onlineCourse.getImage()));
+        }
+        final Image newImage = loadImage(updateOnlineLessonInfoCommand.imageSrc());
+        onlineCourse.setImage(newImage);
     }
 
     private Image loadImage(final String imageSrc) {
